refactor(llm): move known_model_profiles and ModelProfile types to core
Move known_model_profiles module from contracts to core as requested.
Also move ModelProfile, ReasoningEffortConfig, ReasoningLevel types
to core since they're needed by known_model_profiles.

Changes:
- Add model_profile_types.rs to core with the types
- Move known_model_profiles.rs from contracts to core
- Contracts now re-exports the types from core
- Add utoipa dependency to core for ToSchema derives
