+++ crates/everruns-contracts/src/lib.rs
@@ -6,7 +6,6 @@ pub mod agent;
 pub mod capability;
 pub mod common;
 pub mod events;
-pub mod known_model_profiles;
 pub mod llm;
 pub mod message;
 pub mod session;
@@ -16,7 +15,6 @@ pub use agent::*;
 pub use capability::*;
 pub use common::*;
 pub use events::*;
-pub use known_model_profiles::*;
 pub use llm::*;
 pub use message::*;
 pub use session::*;
