+++ crates/everruns-core/src/lib.rs
@@ -12,12 +12,13 @@
 // - ToolRegistry implements ToolExecutor for easy tool management
 // - Error handling distinguishes between user-visible and internal errors
 // - Capabilities provide modular functionality units for composing agent behavior
-// - This crate has NO dependency on storage/contracts - it's purely runtime abstractions
-// - AG-UI events and tool types are defined here as runtime types
+// - This crate has NO dependency on storage - it's purely runtime abstractions
+// - AG-UI events, tool types, and model profile types are defined here
 
-// Runtime types (AG-UI protocol events, tool definitions, capability types)
+// Runtime types (AG-UI protocol events, tool definitions, capability types, model profiles)
 pub mod ag_ui;
 pub mod capability_types;
+pub mod model_profile_types;
 pub mod tool_types;
 
 pub mod atoms;
@@ -40,6 +41,10 @@ pub mod memory;
 pub mod anthropic;
 pub mod openai;
 pub mod provider_factory;
+
+// Known model profiles (hardcoded capabilities for common models)
+// Based on data from https://models.dev/api.json
+pub mod known_model_profiles;
 
 // Re-exports for convenience
 pub use config::AgentConfig;
@@ -95,7 +100,13 @@ pub use ag_ui::{
 // Tool types (runtime types defined in this crate)
 pub use tool_types::{BuiltinTool, ToolCall, ToolDefinition, ToolPolicy, ToolResult};
 
+// Model profile types (runtime types defined in this crate)
+pub use model_profile_types::{ModelProfile, ReasoningEffortConfig, ReasoningLevel};
+
 // Provider factory re-exports
 pub use provider_factory::{create_provider, BoxedLlmProvider, ProviderConfig, ProviderType};
+
+// Known model profiles re-exports
+pub use known_model_profiles::{get_effective_model_profile, get_known_model_profile};
 
 // Note: CapabilityId and CapabilityStatus are re-exported via capabilities module
